backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Rate limiting
slowapi==0.1.9

# File handling
python-magic==0.4.27
pillow==10.1.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
